### 2021-07-10 #3463 @Kevin-Mattheus-Moerman
@m-stclair thanks for this submission. I believe the paper compilation might have been unsuccessful due to the formatting of the affiliations. I attempted a fix which you can review and accept here: https://github.com/MillionConcepts/lhorizon/pull/3).

Also please check the invalid DOI listed :point_up:.

You can call `@whedon generate pdf` here to update the paper.
