### 2021-07-10 #3463 @m-stclair
Thank you for your notes! I have merged your fix and fixed the DOI, and the paper now appears to compile successfully.
