### 2021-07-14 #3463 @arfon
:wave: @malmans2 @AndrewAnnex @steo85it – would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The submission under review here is [_lhorizon: geometry and targeting via JPL Horizons_](https://github.com/MillionConcepts/lhorizon), a Python package for interacting with the JPL Horizons service.
