### 2021-07-14 #3463 @AndrewAnnex
hey @arfon, I would love to be reviewer but I have worked on proposals with/contributed code to @MillionConcepts and @m-stclair, and so I think I have a personal association COI with the company and first author, and this project also lists SpiceyPy as a dependency.

That said the most recent proposal I was listed on with @m-stclair was not funded (IIRC also in 2019). Let me know if you have any questions.
