### 2021-07-14 #3463 @arfon
> hey @arfon, I would love to be reviewer but I have worked on proposals with/contributed code to @MillionConcepts and @m-stclair, and so I think I have a personal association COI with the company and first author, and this project also lists SpiceyPy as a dependency.

Got it, thanks @AndrewAnnex. I'll keep looking for alternative reviewers for now then.
