### 2021-07-15 #3463 @steo85it
Hi @arfon, happy to review this.
-------- Messaggio originale --------Da: Arfon Smith ***@***.***> Data: 14/07/21  6:11 AM  (GMT-05:00) A: openjournals/joss-reviews ***@***.***> Cc: Stefano Bertone ***@***.***>, Mention ***@***.***> Oggetto: Re: [openjournals/joss-reviews] [PRE REVIEW]: lhorizon: geometry and targeting via JPL Horizons (#3463)

👋 @malmans2 @AndrewAnnex @steo85it – would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The submission under review here is lhorizon: geometry and targeting via JPL Horizons, a Python package for interacting with the JPL Horizons service.

—You are receiving this because you were mentioned.Reply to this email directly, view it on GitHub, or unsubscribe.
