### 2021-07-15 #3495 @whedon
[REVIEW]: lhorizon: geometry and targeting via JPL Horizons
**Submitting author:** @m-stclair (<a href="http://orcid.org/0000-0002-7877-3148">Michael St. Clair</a>)
**Repository:** <a href="https://github.com/MillionConcepts/lhorizon" target ="_blank">https://github.com/MillionConcepts/lhorizon</a>
**Version:** 1.0.0
