### 2021-07-16 MillionConcepts/lhorizon#4 @malmans2
Test failure
Hi there,

I've run the tests and I'm getting one failure. I'm using this [environment](https://github.com/MillionConcepts/lhorizon/blob/main/environment.yml).
Here is the log: [pytest.log](https://github.com/MillionConcepts/lhorizon/files/6830705/pytest.log)

On a slightly different note: Have you considered a [github CI workflow](https://docs.github.com/en/actions/guides/building-and-testing-python)?
I've set it up for a few python projects, so I'd be happy to send a PR if that works for you.

https://github.com/openjournals/joss-reviews/issues/3495
