### 2021-07-19 MillionConcepts/lhorizon#4 @m-stclair
Sorry, I wasn't getting notifications from this thread.

Test is not taking the value modulo 360! Thank you, I'll fix that.

I have considered it but have never used one. If you'd like to send a PR, that would be great.
