### 2021-07-19 c86f838
fixing modulo 360 issue in test_targeter_online
