### 2021-07-20 eadf183
setup github workflow, codecov, and add badges
