+++ readme.md
@@ -1,5 +1,8 @@
 # lhorizon
 
+[![CI](https://img.shields.io/github/workflow/status/MillionConcepts/lhorizon/CI?logo=github)](https://github.com/MillionConcepts/lhorizon/action)
+[![codecov](https://codecov.io/gh/MillionConcepts/lhorizon/branch/main/graph/badge.svg)](https://codecov.io/gh/MillionConcepts/lhorizon)
+
 ## introduction
 
 ### purpose
