+++ .github/workflows/ci.yaml
@@ -1,0 +1,46 @@
+name: CI
+
+on:
+  push:
+    branches: "main"
+  pull_request:
+    branches: "*"
+  schedule:
+    - cron: "0 13 * * 1"
+
+jobs:
+  build:
+    name: ${{ matrix.os }} py${{ matrix.python-version }}
+    runs-on: ${{ matrix.os }}
+    strategy:
+      fail-fast: false
+      matrix:
+        os: ["ubuntu-latest", "macos-latest", "windows-latest"]
+        python-version: ["3.9"]
+    steps:
+      - uses: actions/checkout@v2
+      - uses: conda-incubator/setup-miniconda@v2
+        with:
+          environment-file: environment.yml
+          activate-environment: lhorizon
+          auto-update-conda: false
+          python-version: ${{ matrix.python-version }}
+
+      - name: Set up conda environment
+        shell: bash -l {0}
+        run: |
+          python -m pip install -e .
+          conda list
+
+      - name: Run Tests
+        shell: bash -l {0}
+        run: pytest --cov-report term --cov-report xml
+
+      - name: Upload code coverage to Codecov
+        uses: codecov/codecov-action@v1
+        with:
+          file: ./coverage.xml
+          flags: unittests
+          env_vars: OS,PYTHON
+          name: codecov-umbrella
+          fail_ci_if_error: false
