### 2021-07-20 af255bb
Add pytest-cov to environment
