+++ environment.yml
@@ -9,6 +9,7 @@ dependencies:
   - python=3.9
   - pandas
   - pytest
+  - pytest-cov
   - pytest-mock
   - requests
   - spiceypy
