### 2021-07-20 1f0f29a
Update ci.yaml
tweaking pytest-cov settings
