+++ .github/workflows/ci.yaml
@@ -34,13 +34,13 @@ jobs:
 
       - name: Run Tests
         shell: bash -l {0}
-        run: pytest --cov-report term --cov-report xml
+        run: pytest --cov=lhorizon --cov-report xml:.coverage.xml
 
       - name: Upload code coverage to Codecov
-        uses: codecov/codecov-action@v1
+        uses: codecov/codecov-action@v2
         with:
-          file: ./coverage.xml
-          flags: unittests
+          file:.coverage.xml
           env_vars: OS,PYTHON
           name: codecov-umbrella
           fail_ci_if_error: false
+          verbose: true
