### 2021-07-21 MillionConcepts/lhorizon#7 @malmans2
Separate optional from required dependencies
I noticed that a few optional/test dependencies are listed in `install_requires`: https://github.com/MillionConcepts/lhorizon/blob/b7444036f6df973ba9110bc39b88fef02edc84ca/setup.py#L11-L21

I'd remove those from there and add separate entries (e.g., `tests_require` and `extras_require`).

You might want to test both a complete and bare environment in your CI workflow to make sure that everything is working as expected. If you do so, I find this python function quite useful: https://docs.pytest.org/en/6.2.x/reference.html#pytest-importorskip

https://github.com/openjournals/joss-reviews/issues/3495
