### 2021-07-21 MillionConcepts/lhorizon#8 @malmans2
Improve test coverage?
Looks like there are a few modules that have a lower test coverage: https://codecov.io/gh/MillionConcepts/lhorizon/tree/main/lhorizon
Coverage % is not necessarily the best metric though... Do you think the test suite is already complete and robust, or is there scope for improvement?

Particularly, the `handlers` module is not tested at all. Is there a particular reason for that?

https://github.com/openjournals/joss-reviews/issues/3495
