### 2021-07-22 32bc025
tests for handlers; related cleanup and minor refactoring
