### 2021-07-22 MillionConcepts/lhorizon#8 @m-stclair
> > Particularly, the `handlers` module is not tested at all. Is there a particular reason for that?
>
> Probably laziness. The members of `handlers` are basically glue functions...

added reasonably meaningful tests for `handlers`: https://github.com/MillionConcepts/lhorizon/commit/981374968a3ff0bb8a9eeeb196a1c8cc72be4a80
