### 2021-07-25 41fdb19
additional tests for lhorizon_utils, better time, better triangles
