+++ lhorizon/lhorizon_utils.py
@@ -1,10 +1,10 @@
 import datetime as dt
 import pytz
 import re
-from collections.abc import Callable, Sequence
-from functools import reduce
+from collections.abc import Callable, Iterable, Sequence
+from functools import reduce, partial
 from math import floor
-from operator import or_
+from operator import or_, and_, contains
 from typing import Any, Optional, Pattern, Union
 
 import dateutil.parser as dtp
@@ -16,11 +16,14 @@ from numpy.linalg import norm
 
 from lhorizon import config as config
 from lhorizon._type_aliases import Array, Timelike
+from lhorizon.constants import LEAP_SECOND_THRESHOLDS, DT_J2000_TDB
 
 
 def listify(thing: Any) -> list:
     """Always a list, for things that want lists. use with care."""
-    if isinstance(thing, str) or (not isinstance(thing, Sequence)):
+    if isinstance(thing, str) or (
+        (not isinstance(thing, Sequence) and (not isinstance(thing, Iterable)))
+    ):
         return [thing]
     return list(thing)
 
@@ -59,8 +62,8 @@ def hunt_csv(regex: Pattern, body: str) -> list:
     worse than StringIO -> numpy or pandas csv reader in other cases.
     """
     csv_string = re.search(regex, body)[0]
-    if r"\n" in csv_string:
-        lines = csv_string.split(r"\n")
+    if are_in(["\r", "\n"], or_)(csv_string):
+        lines = re.split(r"[\r\n]+", csv_string)
         processed_lines = []
         for line in lines:
             csv_fields = line.split(",")
@@ -71,35 +74,27 @@ def hunt_csv(regex: Pattern, body: str) -> list:
     return [field.strip() for field in csv_fields]
 
 
-LEAP_SECOND_THRESHOLDS = [
-    dt.datetime(1972, 6, 30, tzinfo=dt.timezone.utc),
-    dt.datetime(1972, 12, 31, tzinfo=dt.timezone.utc),
-    dt.datetime(1973, 12, 31, tzinfo=dt.timezone.utc),
-    dt.datetime(1974, 12, 31, tzinfo=dt.timezone.utc),
-    dt.datetime(1975, 12, 31, tzinfo=dt.timezone.utc),
-    dt.datetime(1976, 12, 31, tzinfo=dt.timezone.utc),
-    dt.datetime(1977, 12, 31, tzinfo=dt.timezone.utc),
-    dt.datetime(1978, 12, 31, tzinfo=dt.timezone.utc),
-    dt.datetime(1979, 12, 31, tzinfo=dt.timezone.utc),
-    dt.datetime(1981, 6, 30, tzinfo=dt.timezone.utc),
-    dt.datetime(1982, 6, 30, tzinfo=dt.timezone.utc),
-    dt.datetime(1983, 6, 30, tzinfo=dt.timezone.utc),
-    dt.datetime(1985, 6, 30, tzinfo=dt.timezone.utc),
-    dt.datetime(1987, 12, 31, tzinfo=dt.timezone.utc),
-    dt.datetime(1989, 12, 31, tzinfo=dt.timezone.utc),
-    dt.datetime(1990, 12, 31, tzinfo=dt.timezone.utc),
-    dt.datetime(1992, 6, 30, tzinfo=dt.timezone.utc),
-    dt.datetime(1993, 6, 30, tzinfo=dt.timezone.utc),
-    dt.datetime(1994, 6, 30, tzinfo=dt.timezone.utc),
-    dt.datetime(1995, 12, 31, tzinfo=dt.timezone.utc),
[…]

[…]
