### 2021-07-25 8864e37
adding additional targeter edge case testing, removing some cruft
