+++ lhorizon/targeter_utils.py
@@ -1,29 +1,10 @@
-import warnings
 from itertools import repeat
-from typing import Mapping, Callable, Sequence, Any
+from typing import Sequence
 
 import numpy as np
 import spiceypy as spice
 
 from lhorizon._type_aliases import Array
-
-
-def find_intersections(
-    solutions: Mapping[Any, Callable],
-    ray_direction: Array,
-    target_center: Array,
-) -> dict:
-    """
-    find intersections between ray_direction and target_center given a mapping
-    of functions (like output of `solutions.make_ray_sphere_lambdas`)
-    """
-    # suppress irrelevant warnings about imaginary values
-    with warnings.catch_warnings:
-        warnings.simplefilter("ignore")
-        return {
-            coordinate: solution(*ray_direction, *target_center)
-            for coordinate, solution in solutions.items()
-        }
 
 
 def array_reference_shift(
