### 2021-07-25 3cc0bb2
attempt adding minimal environment tests to workflow
