+++ .github/workflows/ci.yaml
@@ -21,6 +21,24 @@ jobs:
       - uses: actions/checkout@v2
       - uses: conda-incubator/setup-miniconda@v2
         with:
+          environment-file: .minimal-environment.yml
+          activate-environment: lhorizon-minimal
+          auto-update-conda: false
+          python-version: ${{ matrix.python-version }}
+
+      - name: Set up minimal conda environment
+        shell: bash -l {0}
+        run: |
+          python -m pip install -e .
+          conda list
+
+      - name: Run Tests with minimal environment
+        shell: bash -l {0}
+        run: pytest
+
+      - uses: actions/checkout@v2
+      - uses: conda-incubator/setup-miniconda@v2
+        with:
           environment-file: environment.yml
           activate-environment: lhorizon
           auto-update-conda: false
