### 2021-07-25 MillionConcepts/lhorizon#8 @m-stclair
improved test coverage, edge handling, cruft removal for `lhorizon_utils`, `target`, and `targeter_utils`: 41fdb19ae1f5a83373787058e81562d84619446f,
8864e37ff5063d8138d7159d57b08f8e69777d70
