### 2021-07-25 MillionConcepts/lhorizon#7 @m-stclair
separate optional from required dependencies: a8668e6aa9465d4fbb307eab9b53954f868c3f0c
add tests with minimal environment: 3cc0bb2c4ed86af44f20d9071f017603e11ea989
