### 2021-08-09 MillionConcepts/lhorizon#9 @malmans2
List all required dependencies
I would list all required dependencies here: https://github.com/MillionConcepts/lhorizon#dependencies-and-requirements

https://github.com/openjournals/joss-reviews/issues/3495
