### 2021-08-09 MillionConcepts/lhorizon#10 @malmans2
Don't use 1st person in README
There are two authors on the paper and it looks like there are two maintainers.
In the README, I wouldn't use 1st person (e.g., "I wrote"). Either use "we" or the github handle if you prefer to be more specific.

https://github.com/openjournals/joss-reviews/issues/3495https://github.com/openjournals/joss-reviews/issues/3495
