### 2021-08-09 5f555fd
Use nbviewer link
I suggest to use the nbviewer link for notebooks, just in case you'll add bigger notebooks in the future.

https://github.com/openjournals/joss-reviews/issues/3495
