### 2021-08-09 MillionConcepts/lhorizon#12 @malmans2
Set up binder?
You can easily make the example notebooks interactive using [binder](https://mybinder.org/).
It would be easier for new user to explore lhorizon, especially because the notebooks are stored without outputs.
I'd be happy to open a PR.

https://github.com/openjournals/joss-reviews/issues/3495
