### 2021-08-09 MillionConcepts/lhorizon#8 @malmans2
Thanks! See #13

I would also add instructions on how to locally run the tests in the README.
