### 2021-08-09 MillionConcepts/lhorizon#14 @malmans2
Use sphinx to generate docs?
If I'm understanding correctly, `api.md` is not automatically inferred from the actual code but was manually written.
I think it would be more robust to only use docstrings in the code, this way the api in the documentation is always up to date.
I suggest to look into [sphinx](https://www.sphinx-doc.org/en/master/), particularly the docstring extensions: [google style](https://www.sphinx-doc.org/en/master/usage/extensions/example_google.html), [numpy style](https://www.sphinx-doc.org/en/master/usage/extensions/example_numpy.html#example-numpy)

https://github.com/openjournals/joss-reviews/issues/3495
