### 2021-08-09 b87f568
Update .minimal-environment.yml
