+++ .minimal-environment.yml
@@ -9,4 +9,5 @@ dependencies:
   - pandas
   - pytest
   - pytest-mock
+  - pytest-cov
   - requests
