### 2021-08-09 MillionConcepts/lhorizon#15 @malmans2
OBSERVER table quantity codes
I'm wondering if it's worth adding a function/method to retrieve the footnote at the end of [this notebook](https://github.com/MillionConcepts/lhorizon/blob/main/lhorizon/examples/mars_sun_angle.ipynb)?

https://github.com/openjournals/joss-reviews/issues/3495
