### 2021-08-11 MillionConcepts/lhorizon#14 @m-stclair
`api.md` is generated using `mkdocs` -- i prefer markdown docstrings to google or numpy-style docstrings for this project. I am, however, regenerating it only manually -- do you think it would be preferable to add a CI hook to rebuild the docs?
