### 2021-08-11 MillionConcepts/lhorizon#9 @m-stclair
addressed in https://github.com/MillionConcepts/lhorizon/commit/27fb3ab5fda3745101ca6d75e430270b1096bf57
