### 2021-08-11 MillionConcepts/lhorizon#10 @m-stclair
addressed in https://github.com/MillionConcepts/lhorizon/commit/9acc043374f0f81889f4f02e0a4e4db082dc0f03
