### 2021-08-11 MillionConcepts/lhorizon#15 @m-stclair
I would _like_ to do this. Unfortunately, as far as I can tell, there is no way to retrieve this concise table of quantity codes from the query interface. I can retrieve a formal prose description of every quantity value, but this is about 20 pages long and would not serve the same purpose. I _think_ the best I could do is add a reference to the section of the web documentation this table is taken from. (I suppose I could also scrape the text from the web documentation, but this does not seem very useful or guaranteed to be stable.)
