### 2021-08-11 MillionConcepts/lhorizon#15 @m-stclair
oh, or are you perhaps suggesting that I just squirrel it away as a string literal somewhere in the module and add a function/method to retrieve it?
