### 2021-08-11 MillionConcepts/lhorizon#14 @malmans2
> do you think it would be preferable to add a CI hook to rebuild the docs?

Yes, I think that would be ideal. You could either add a github action that rebuilds the API and automatically submits a commit/PR, or you could add a test function that rebuilds the API and checks whether it is identical to the existing one. If not, pytest will report a failure and you'll know that you have to rebuild the API.
