### 2021-08-12 MillionConcepts/lhorizon#15 @malmans2
> oh, or are you perhaps suggesting that I just squirrel it away as a string literal somewhere in the module and add a function/method to retrieve it?

If there's no easy way to retrieve it from the description, yes I think it would work as long as the table is not going to change. Otherwise, maybe just a clear link in the documentation to the long description would be helpful. I'll leave it up to you.
