### 2021-08-12 MillionConcepts/lhorizon#17 @malmans2
Publish to PyPI and/or conda-forge?
You could publish the package on PyPI and conda-forge.
It's quite easy to set up a GH workflow that will publish every time a new tag is created.

The basic workflow is described [here](https://packaging.python.org/guides/publishing-package-distribution-releases-using-github-actions-ci-cd-workflows/), but there are many tutorials online.

https://github.com/openjournals/joss-reviews/issues/3495
