### 2021-08-12 4516a32
adding functions to retrieve observer quantity code table
