### 2021-08-12 MillionConcepts/lhorizon#15 @m-stclair
the situation is actually somewhat worse than I thought. the table provided in the documentation is explicitly noted as subject to change at any time and is already out of date, meaning that the only way to access an up-to-date table of quantity codes is via telnet or the interactive web CGI (which is different from the batch CGI used by lhorizon and would require spoofing browser interaction to access programmatically)...I thought a little more about the problem and found a way to trick the telnet endpoint into printing the table. this is probably a little fragile, but better than forcing users to perform multiple interaction steps on external resources to have a correct list of options. see https://github.com/MillionConcepts/lhorizon/commit/166f7291ef9b09c14bebc9b396087a8d0be6d7a6
