### 2021-08-12 6be98e0
move example notebooks to root directory
