### 2021-08-12 MillionConcepts/lhorizon#14 @m-stclair
there's an apparent regression in `pydoc-markdown` preventing me from automating this, so I'm putting a pin in this for now. if the issue isn't resolved in a couple of weeks, I'll switch to some other documentation-generating system.
