### 2021-08-12 9058429
update docs, add test instructions
