+++ docs/api.md
@@ -12,6 +12,7 @@
   * [lambdify\_system](#solutions.lambdify_system)
   * [make\_ray\_sphere\_lambdas](#solutions.make_ray_sphere_lambdas)
 * [config](#config)
+* [dscratch](#dscratch)
 * [\_request\_formatters](#_request_formatters)
   * [format\_geodetic\_origin](#_request_formatters.format_geodetic_origin)
   * [format\_geodetic\_target](#_request_formatters.format_geodetic_target)
@@ -28,7 +29,6 @@
     * [\_\_str\_\_](#base.LHorizon.__str__)
     * [\_\_repr\_\_](#base.LHorizon.__repr__)
 * [targeter\_utils](#targeter_utils)
-  * [find\_intersections](#targeter_utils.find_intersections)
   * [array\_reference\_shift](#targeter_utils.array_reference_shift)
 * [target](#target)
   * [Targeter](#target.Targeter)
@@ -44,25 +44,32 @@
   * [query\_all\_lhorizons](#handlers.query_all_lhorizons)
   * [list\_sites](#handlers.list_sites)
   * [list\_majorbodies](#handlers.list_majorbodies)
+  * [get\_observer\_quantity\_codes](#handlers.get_observer_quantity_codes)
 * [lhorizon\_utils](#lhorizon_utils)
   * [listify](#lhorizon_utils.listify)
   * [snorm](#lhorizon_utils.snorm)
   * [hunt\_csv](#lhorizon_utils.hunt_csv)
+  * [are\_in](#lhorizon_utils.are_in)
+  * [is\_it](#lhorizon_utils.is_it)
+  * [utc\_to\_tt\_offset](#lhorizon_utils.utc_to_tt_offset)
+  * [utc\_to\_tdb](#lhorizon_utils.utc_to_tdb)
   * [dt\_to\_jd](#lhorizon_utils.dt_to_jd)
   * [numeric\_columns](#lhorizon_utils.numeric_columns)
   * [produce\_jd\_series](#lhorizon_utils.produce_jd_series)
   * [time\_series\_to\_et](#lhorizon_utils.time_series_to_et)
-  * [is\_it](#lhorizon_utils.is_it)
   * [sph2cart](#lhorizon_utils.sph2cart)
   * [cart2sph](#lhorizon_utils.cart2sph)
   * [hats](#lhorizon_utils.hats)
   * [make\_raveled\_meshgrid](#lhorizon_utils.make_raveled_meshgrid)
   * [default\_lhorizon\_session](#lhorizon_utils.default_lhorizon_session)
+  * [perform\_telnet\_exchange](#lhorizon_utils.perform_telnet_exchange)
+  * [have\_telnet\_conversation](#lhorizon_utils.have_telnet_conversation)
 * [kernels](#kernels)
 * [kernels.load](#kernels.load)
   * [load\_metakernel](#kernels.load.load_metakernel)
 
 <a name="_response_parsers"></a>
+
 # Module \_response\_parsers
 
 helper functions for parsing response text from the JPL Horizons CGI.
@@ -70,6 +77,7 @@ these functions are intended to be called by LHorizon methods and should
 generally not be called directly.
 
 <a name="_response_parsers.make_lhorizon_dataframe"></a>
+
 #### make\_lhorizon\_dataframe
 
 ```python
@@ -79,6 +87,7 @@ def make_lhorizon_dataframe(jpl_response: str, topocentric_target: bool = False)
 make a DataFrame from Horizons CGI response text.
 
 <a name="_response_parsers.clean_up_vectors_series"></a>
+
 #### clean\_up\_vectors\_series
 
 ```python
[…]
