+++ lhorizon/__init__.py
@@ -1,3 +1,4 @@
 from lhorizon.base import LHorizon
 
-__version__ = "1.0"
+__version__ = "1.0.1a.1"
+name = "lhorizon"
