+++ .github/workflows/pypi.yaml
@@ -1,0 +1,36 @@
+name: publish to pypi
+on: push
+jobs:
+  build-and-publish:
+    name: build and publish
+    runs-on: "ubuntu-latest"
+    steps:
+      - uses: actions/checkout@v2
+      - uses: conda-incubator/setup-miniconda@v2
+        with:
+          environment-file: environment.yml
+          activate-environment: lhorizon
+          auto-update-conda: false
+          python-version: 3.9
+      - name: Set up conda environment
+        shell: bash -l {0}
+        run: |
+          python -m pip install -e .
+          conda list
+      - name: install build
+        run: conda install -n lhorizon build
+      - name: build binary & source
+        shell: bash -l {0}
+        run: |
+          conda activate lhorizon
+          python -m build --sdist --wheel --outdir dist/
+      - name: publish to TestPyPI
+        uses: pypa/gh-action-pypi-publish@master
+        with:
+          password: ${{ secrets.TEST_PYPI_API_TOKEN }}
+          repository_url: https://test.pypi.org/legacy/
+      - name: Publish distribution to PyPI
+        if: startsWith(github.ref, 'refs/tags')
+        uses: pypa/gh-action-pypi-publish@master
+        with:
+          password: ${{ secrets.PYPI_API_TOKEN }}
