### 2021-08-12 e9e0eee
switch publication to manual
