+++ lhorizon/__init__.py
@@ -1,4 +1,4 @@
 from lhorizon.base import LHorizon
 
-__version__ = "1.0.1a.1"
+__version__ = "1.0.1a.2"
 name = "lhorizon"
