+++ .github/workflows/pypi.yaml
@@ -1,5 +1,6 @@
 name: publish to pypi
-on: push
+on: workflow_dispatch
+
 jobs:
   build-and-publish:
     name: build and publish
