### 2021-08-13 MillionConcepts/lhorizon#17 @m-stclair
I have added actions for publishing to TestPyPi and PyPi, and `lhorizon` is [now available on PyPi](https://pypi.org/project/lhorizon/).

I have also created a `conda-forge` recipe for `lhorizon` and [submitted a PR to staged-recipes](https://github.com/conda-forge/staged-recipes/pull/15898). Leaving this issue open pending resolution of that PR.
