### 2021-08-13 MillionConcepts/lhorizon#17 @malmans2
Nice!
I think it's now better to suggest `pip/conda install lhorizon` in the installation section.

I would probably do something consistent for binder as well.
E.g., replace
https://github.com/MillionConcepts/lhorizon/blob/a25f71435ea8b9b5de5942c735163789cd7eaadb/.binder/postBuild#L3
with
`conda install -c conda-forge pip # or just add pip to environment.yml`
`pip install -e .`
