+++ .github/workflows/ubuntu_tests.yaml
@@ -1,4 +1,4 @@
-name: CI
+name: ubuntu full env tests
 
 on:
   push:
@@ -15,9 +15,9 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        os: ["ubuntu-latest", "macos-latest", "windows-latest"]
+        os: ["ubuntu-latest"]
         python-version: ["3.9"]
-        environment-file: ["environment.yml", ".minimal-environment.yml"]
+        environment-file: ["environment.yml"]
     steps:
       - uses: actions/checkout@v2
       - uses: conda-incubator/setup-miniconda@v2
