+++ setup.py
@@ -2,7 +2,7 @@ from setuptools import setup, find_packages
 
 setup(
     name="lhorizon",
-    version="1.0.1a.2",
+    version="1.1.0",
     description="lhorizon helps you find where things are in the solar system",
     url="https://github.com/millionconcepts/lhorizon.git",
     author="Million Concepts",
