+++ lhorizon/__init__.py
@@ -1,4 +1,4 @@
 from lhorizon.base import LHorizon
 
-__version__ = "1.0.1a.2"
+__version__ = "1.1.0"
 name = "lhorizon"
