+++ .github/workflows/windows_tests.yaml
@@ -35,13 +35,4 @@ jobs:
 
       - name: Run Tests
         shell: bash -l {0}
-        run: pytest --cov-config=.coveragerc --cov=lhorizon --cov-report xml:.coverage.xml
-
-      - name: Upload code coverage to Codecov
-        uses: codecov/codecov-action@v2
-        with:
-          file: .coverage.xml
-          env_vars: OS,PYTHON
-          name: codecov-umbrella
-          fail_ci_if_error: false
-          verbose: true
+        run: pytest
