### 2021-08-13 MillionConcepts/lhorizon#17 @m-stclair
I’ve made install docs consistent and updated binder, and the package is up on c-f! https://anaconda.org/conda-forge/lhorizon
