### 2021-08-14 MillionConcepts/lhorizon#19 @malmans2
Add notebook illustrating performance claims
I would add in the documentation a notebook that illustrates the advantages of using `lhorizon` rather than `astroquery` and supports the performance improvements discussed in the paper (line 27-43).

https://github.com/openjournals/joss-reviews/issues/3495
