+++ lhorizon/tests/data/test_cases.py
@@ -18,20 +18,20 @@ TEST_CASES = {
             "epochs": {
                 "start": "1959-01-01T00:00:00",
                 "stop": "1959-01-02T01:00:00",
-                "step": "10m",
+                "step": "30m",
             },
         },
         "request_url": "https://ssd.jpl.nasa.gov/horizons_batch.cgi?batch=1"
-                       "&TABLE_TYPE=OBSERVER&QUANTITIES=%271%2C2%2C4%2C13"
-                       "%2C14%2C15%2C17%2C20%2C45%27&COMMAND=%22g%3A-9.46"
-                       "%2C40.75%2C0%40499%22&SOLAR_ELONG=%220%2C180%22"
-                       "&LHA_CUTOFF=0&CSV_FORMAT=YES&CAL_FORMAT=BOTH"
-                       "&ANG_FORMAT=DEG&APPARENT=AIRLESS&REF_SYSTEM=J2000"
-                       "&EXTRA_PREC=NO&CENTER=coord%40399&COORD_TYPE"
-                       "=GEODETIC&SITE_COORD=%27-243.460000%2C33.800000%2C0"
-                       ".000000%27&START_TIME=%221959-01-01T00%3A00%3A00%22"
-                       "&STOP_TIME=%221959-01-02T01%3A00%3A00%22&STEP_SIZE"
-                       "=%2210m%22&SKIP_DAYLT=NO",
+        "&TABLE_TYPE=OBSERVER&QUANTITIES=%271%2C2%2C4%2C13"
+        "%2C14%2C15%2C17%2C20%2C45%27&COMMAND=%22g%3A-9.46"
+        "%2C40.75%2C0%40499%22&SOLAR_ELONG=%220%2C180%22"
+        "&LHA_CUTOFF=0&CSV_FORMAT=YES&CAL_FORMAT=BOTH"
+        "&ANG_FORMAT=DEG&APPARENT=AIRLESS&REF_SYSTEM=J2000"
+        "&EXTRA_PREC=NO&CENTER=coord%40399&COORD_TYPE"
+        "=GEODETIC&SITE_COORD=%27-243.460000%2C33.800000%2C0"
+        ".000000%27&START_TIME=%221959-01-01T00%3A00%3A00%22"
+        "&STOP_TIME=%221959-01-02T01%3A00%3A00%22&STEP_SIZE"
+        "=%2230m%22&SKIP_DAYLT=NO",
         "data_path": str(
             Path(Path(__file__).parent, "CYDONIA_PALM_SPRINGS_1959_TOPO")
         ),
@@ -42,12 +42,12 @@ TEST_CASES = {
             "origin": "5@399",
         },
         "request_url": "https://ssd.jpl.nasa.gov/horizons_batch.cgi?batch=1"
-                       "&TABLE_TYPE=OBSERVER&QUANTITIES=%271%2C2%2C4%2C13"
-                       "%2C14%2C15%2C17%2C20%2C45%27&COMMAND=%22301%22"
-                       "&SOLAR_ELONG=%220%2C180%22&LHA_CUTOFF=0&CSV_FORMAT"
-                       "=YES&CAL_FORMAT=BOTH&ANG_FORMAT=DEG&APPARENT=AIRLESS"
-                       "&REF_SYSTEM=J2000&EXTRA_PREC=NO&CENTER=%275%40399%27"
-                       "&TLIST=",
+        "&TABLE_TYPE=OBSERVER&QUANTITIES=%271%2C2%2C4%2C13"
+        "%2C14%2C15%2C17%2C20%2C45%27&COMMAND=%22301%22"
[…]
