+++ lhorizon/_response_parsers.py
@@ -12,10 +12,16 @@ import numpy as np
 import pandas as pd
 from dateutil import parser as dtp
 
-from lhorizon.config import TABLE_PATTERNS
+from lhorizon.config import TABLE_PATTERNS, VISIBILITY_FLAG_NAMES
 from lhorizon.constants import AU_TO_M
-from lhorizon.lhorizon_utils import hunt_csv
+from lhorizon.lhorizon_utils import hunt_csv, \
+    convert_horizons_date_spec_to_strftime
 from lhorizon._type_aliases import Array
+
+
+HORIZON_COLUMN_SEARCH = re.compile(r"(Date|JDTDB).*(?=\n\*+)")
+HORIZON_DATA_SEARCH = re.compile(r"\$\$SOE\n(.*)\$\$EOE", re.DOTALL)
+GEODETIC_SEARCH = re.compile(r"(?<=Target geodetic : )\d.*(?= {)")
 
 
 def make_lhorizon_dataframe(
@@ -26,14 +32,14 @@ def make_lhorizon_dataframe(
     """
     # delimiters for column and data sections
     # 'JDTDB' begins the vectors columns; 'Date' begins the observer columns
-    horizon_column_search = re.compile(r"(Date|JDTDB).*(?=\n\*+)")
-    horizon_data_search = re.compile(
-        r"(?<=\$\$SOE\n)(.|\n)*?(\d\d\d\d(.|\n)*)(?=\$\$EOE)"
-    )
+
     # grab these sections and write them into a string buffer
     try:
-        columns = re.search(horizon_column_search, jpl_response)[0]
-        data = re.search(horizon_data_search, jpl_response).group(2)
+        # find bounds of column / data in response & strip spaces from columns
+        columns = re.search(
+            HORIZON_COLUMN_SEARCH, jpl_response
+        )[0].replace(" ", "")
+        data = re.search(HORIZON_DATA_SEARCH, jpl_response).group(1)
     except TypeError:
         raise ValueError(
             "Horizons didn't return a table of data or it couldn't be parsed "
@@ -43,29 +49,41 @@ def make_lhorizon_dataframe(
     data_buffer.write(columns + "\n" + data)
     data_buffer.seek(0)
     # read this buffer as csv
-    horizon_dataframe = pd.read_csv(data_buffer, sep=" *, *", engine="python")
-    # name the unlabeled flag columns
-    horizon_dataframe.rename(
-        mapper={
-            "Unnamed: 2": "solar_presence",
-            "Unnamed: 3": "lunar_presence",
-            "Unnamed: 4": "nearside_flag",
-            "Unnamed: 5": "illumination_flag",
-        },
-        axis=1,
-        inplace=True,
-    )
-    # add the target's geodetic coordinates if desired
+    horizon_dataframe = pd.read_csv(data_buffer, sep=",", engine="c", low_memory=False)
+    # horizons ends lines w/commas, so pandas creates an empty trailing column
+    horizon_dataframe = horizon_dataframe.iloc[:, :-1]
+    horizon_dataframe = clean_visibility_flags(horizon_dataframe)
+    # if appropriate, add target's geodetic coordinates
     if topocentric_target:
-        horizon_target_search = re.compile(
-            r"(?<=Target geodetic : )\d.*(?= {)"
-        )
-        target_geodetic_coords = hunt_csv(horizon_target_search, jpl_response)
+        target_geodetic_coords = hunt_csv(GEODETIC_SEARCH, jpl_response)
         horizon_dataframe["geo_lon"] = target_geodetic_coords[0]
         horizon_dataframe["geo_lat"] = target_geodetic_coords[1]
         horizon_dataframe["geo_el"] = target_geodetic_coords[2]
-    # drop empty columns and return
[…]
