### 2021-08-15 45bd3ee
tweaks to paper text and bibliography
