+++ docs/paper.bib
@@ -1,10 +1,16 @@
 % Encoding: UTF-8
 
-@InProceedings{Giorgini_2015,
-  author    = {Giorgini, J.D.},
-  booktitle = {IAU General Assembly},
-  title     = {Status of the JPL horizons ephemeris system.},
-  year      = {2015},
+@INPROCEEDINGS{2015IAUGA..2256293G,
+       author = {{Giorgini}, Jon D.},
+        title = "{Status of the JPL Horizons Ephemeris System}",
+    booktitle = {IAU General Assembly},
+         year = 2015,
+       volume = {29},
+        month = aug,
+          eid = {2256293},
+        pages = {2256293},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2015IAUGA..2256293G},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
 @Article{Ginsburg_2019,
@@ -37,7 +43,7 @@
   number    = {46},
   pages     = {2050},
   volume    = {5},
-  doi       = {doi.org/10.21105/joss.02050},
+  doi       = {10.21105/joss.02050},
   publisher = {The Open Journal},
 }
 
@@ -99,4 +105,38 @@
   title        = {Skyfield: High precision research-grade positions for planets and Earth satellites generator},
   year         = {2019},
   url          = {https://ui.adsabs.harvard.edu/abs/2019ascl.soft07024R/abstract},
+}
+
+@Article{harris2020array,
+ title         = {Array programming with {NumPy}},
+ author        = {Charles R. Harris and K. Jarrod Millman and St{\'{e}}fan J.
+                 van der Walt and Ralf Gommers and Pauli Virtanen and David
+                 Cournapeau and Eric Wieser and Julian Taylor and Sebastian
+                 Berg and Nathaniel J. Smith and Robert Kern and Matti Picus
+                 and Stephan Hoyer and Marten H. van Kerkwijk and Matthew
+                 Brett and Allan Haldane and Jaime Fern{\'{a}}ndez del
+                 R{\'{i}}o and Mark Wiebe and Pearu Peterson and Pierre
+                 G{\'{e}}rard-Marchant and Kevin Sheppard and Tyler Reddy and
+                 Warren Weckesser and Hameer Abbasi and Christoph Gohlke and
+                 Travis E. Oliphant},
+ year          = {2020},
+ month         = sep,
+ journal       = {Nature},
+ volume        = {585},
+ number        = {7825},
+ pages         = {357--362},
+ doi           = {10.1038/s41586-020-2649-2},
+ publisher     = {Springer Science and Business Media {LLC}},
+ url           = {https://doi.org/10.1038/s41586-020-2649-2}
+}
+
+@software{reback2020pandas,
+    author       = {The pandas development team},
+    title        = {pandas-dev/pandas: Pandas},
+    month        = feb,
+    year         = 2020,
+    publisher    = {Zenodo},
+    version      = {latest},
+    doi          = {10.5281/zenodo.3509134},
+    url          = {https://doi.org/10.5281/zenodo.3509134}
 }
