+++ benchmark/bench-environment.yml
@@ -1,0 +1,17 @@
+name: lh-bench
+channels:
+  - conda-forge
+  - defaults
+dependencies:
+  - more-itertools
+  - numpy
+  - python=3.9
+  - pandas
+  - pytest
+  - pytest-mock
+  - pytest-cov
+  - requests
+  - astroquery
+  - pympler
+  - memory-profiler
+
