### 2021-08-15 MillionConcepts/lhorizon#19 @m-stclair
Thanks for requesting this...there have been some significant backend improvements to `astropy` tables since I last compared these packages, and I have to walk my claim back to 10x rather than 10-100x. I have revised relevant text to reflect the current situation and avoid unfairly maligning `astropy` tables. (Also, while benchmarking, I found a couple of parser bugs my tests were not adequately covering, so, great.)

I've added some comparison benchmarks in the 'benchmark' directory -- they of course require bringing in additional dependencies, so I haven't added them to the main examples directory. Also, the use cases where the performance improvements are important involve pretty big chunks of geometry data, Horizons' serverside response time is so variable that you cannot do apples-to-apples comparisons with live queries, and I didn't want to add an additional 100 MB to the repository size, so I've stashed the sample data needed for these benchmarks on Google Drive. (Link is in the benchmark Notebook.)

Also, I would like to make sure that it is clear that, for many users, the reasons to use `lhorizon` over `jplhorizons` would be functionality and interface rather than performance. (And conversely, users who don't need `lhorizon`'s features and want tight integration with the `astropy` ecosystem would probably prefer to use `jplhorizons`.) I've tried to make these points a little clearer in the paper.
