### 2021-08-27 MillionConcepts/lhorizon#21 @steo85it
Comments to paper review
I applied most of my suggestions directly to the text in PR #20  (I hope that's convenient for you, else let me know what you prefer). The structure and goals are clear and the article is generally well suited to present the software. However, here are some additional open remarks/questions:
- [x] **_Horizons_ is the only system in the world that provides...**: not really, other interfaces exist, such as https://ssp.imcce.fr/forms/ephemeris (and others under https://ssp.imcce.fr/forms ) or http://iaaras.ru/en/dept/ephemeris/online/ . The statement could be "weakened" (here and else in the text, e.g., "_Horizons_ is one of the main providers of" or "among few providers of ... in the world.") to make it a bit fairer to other groups.
- [x] **"_Horizons_ is high-precision"**: what does "high precision" mean?
- [x] **"has the capability to produce corrected positions"**: what does "corrected" mean here? "corrected" of what effects/errors/etc?
- [x] **"had experienced bitrot"** : this is slang and at least it's not familiar to me
- [x] **"JPL is the most authoritative producer of solar system ephemerides"**: again, this is arbitrary and should be weakened to "one of the few", "one of the most", etc... Several institutes in the world provide ephemerides of comparable accuracy and relevance, notably JPL produces DE, IMCCE produces INPOP (the french/european ephemeris), and the RAS produces EPM (Russian ephemeris). At least the first 2 are regularly used in the framework of space missions.
- [x] **"“simple” quantities like the solar angle"**: which angle? azimuth? elevation? maybe "solar angleS" would easily solve this.
- [x] **"also collecting the appropriate data files, called “kernels.”**: I know by experience that this is often tricky, but wouldn't your examples also get into the same trouble when calling spiceypy (e.g., for reference frames transformations)? If you found a way to automatically load dependencies for the high-level kernels, this should definitely be highlighted more (it would be awesome!!).
- [x] **"automatically-generated kernels"**: what do you mean here?
- [x] **"up to tens of millions of kernels"**: did you really count them? All version summed together? :) Anyway, the sentence was extremely redundant.
- [x] as a general remark, I think the paragraph about SPICE was way too harsh and colloquial. Also, it's hard to defend subjective statement such as "extremely high barrier" or "is difficult for planetary scientists and specialized engineers" (it's not, after you get accustomed to it, as for most software packages). I get your point (mostly), but the language needed amendments (mine in this PR is just a suggestion).
- [x] **"it cover more sites"**: what does this mean? which "sites"?
- [x] **"geometry data"**: it would be good to define this somewhere (rather early in the text), as it's somehow the focus and might be unclear to general readers/users. Also, I'm not sure whether "geometry data" or "geometric data" or "geometry" (e.g., of an observation) would be the correct terminology.
- [x] **Other Related Work**: references to the (very much active) CALCEPH library (https://www.imcce.fr/inpop/calceph/ for C, fortran, python, Matlab interfaces) is missing.
- [x] **Statement of Need**: as a potential user reading this, I would find it useful to have some examples of applications (such as a brief summary of the jupiter notebooks you prepared). This would be a good place to briefly present and reference them.
- [x] well, include suggestions (or similar changes) as included in PR #20
- [x] consistently update README.md
