### 2021-08-27 MillionConcepts/lhorizon#23 @steo85it
examples/target_example.ipynb
This is an interesting example of application of your package and functions!
However, am I right to say that "In 8" and the whole Part 2 fully rely on SPICE/spicey-py?  You correctly state at some point that "additional kernels might be needed for different applications", but still it is not clearly stated that in order to get those functionalities (reference frame changes) one needs to load, use and understand spice kernels (which is fine, but also an issue that you extensively claim to avoid in your paper).

openjournals/joss-reviews#3495
