### 2021-08-27 MillionConcepts/lhorizon#23 @steo85it
Also, minor point: what do you mean by `earth mean equator and equinox of J2000, **often called "J2000" for historical reasons**`? It's simply called J2000 because it refers to the equator at 01/01/2000 12:00:00.
