### 2021-08-27 MillionConcepts/lhorizon#19 @steo85it
Love this! Unfortunately the notebook still fails in binder (at least for me, astropy seems to be missing)
![image](https://user-images.githubusercontent.com/29377066/131192018-79e354cf-b746-4424-a305-694fd80b2ed5.png)

Also, I would suggest to add some `assert` checks among calls results, to make sure/prove that on top of a better timing, one also gets the same results.
