### 2021-08-27 MillionConcepts/lhorizon#24 @steo85it
do tests verify results against external sources (e.g., an Horizons dump)?
As from the title, are the (very extensive and working well) tests provided with the code only checking internal functionalities and consistency, or are they also verifying that results by the package are consistent with output by the `Horizons` platform (or by `spiceypy` or `astropy` or whatever else)?

If yes, I would suggest to:
- [x] add some brief description to routines and functions in `lhorizon/tests`, as that's completely missing at present

If not, then:
- [ ] please add one or more tests comparing your results for a particular application (some vector/range computation between features/bodies) to equivalent information retrieved from other sources (a dump of Horizons output to csv, a spiceypy/astropy call).

https://github.com/openjournals/joss-reviews/issues/3495
