### 2021-08-31 MillionConcepts/lhorizon#19 @malmans2
Sorry about the delay, I was on leave.
I have a couple of suggestions:

- [x]  I think it would be better to rename [bench_readme.md](https://github.com/MillionConcepts/lhorizon/blob/main/benchmark/bench_readme.md) to just README.md (I'm not 100% sure, but I think in this way GitHub will show the README in the benchmark directory without having to click on it)
- [x] Use step by step instructions in the readme, including terminal commands such as how to grab the files (e.g., wget/curl), how to install the benchmark environment, and how to install lhorizon in the benchmark environment
- [x] Consider storing the benchmark files in a different repository/cloud (I believe they are currently in a google drive). E.g., you could use [git-lfs](https://git-lfs.github.com/) and store the file in this GitHub repo, or you could use zenodo.
- [x] I think there is a typo in the environment file. ~`memory-profiler`~ should be `memory_profiler` https://github.com/MillionConcepts/lhorizon/blob/eee7be4ee9e14f72de643bd690163cb0640a657c/benchmark/bench-environment.yml#L16
- [x] I would add `jupyter` in the environment, as it is consistent with the environment file provided in the root directory
- [x] Although the variable names are self-describing, I would make more obvious which cell is benchmarking lhorizon vs jplhorizon (e.g., add markdowns, comments, or prints)
