+++ docs/paper.md
@@ -21,88 +21,130 @@ bibliography: paper.bib
 
 # Summary
 
-`lhorizon` helps you locate natural and artificial bodies and features in the Solar System. It is built around a thick Python wrapper for the Jet 
-Propulsion Laboratory (JPL) Solar System Dynamics Group  _Horizons_ service [@Giorgini_2015]. _Horizons_ is the only 
-system in the world that provides ready-to-go, no-assembly-required access to geometry data for almost every object in 
-the Solar System.  Other interfaces to _Horizons_ exist (see 'Other Related Work' below for several examples), but 
-`lhorizon` is particularly optimized for stability and time-to-value for large queries and for processing results with 
+`lhorizon` helps you locate natural and artificial bodies and features in the
+Solar System. It is built around a thick Python wrapper for the Jet Propulsion
+Laboratory (JPL) Solar System Dynamics Group  _Horizons_ service
+[@Giorgini_2015]. _Horizons_ is the one of the only providers of ready-to-go,
+no-assembly-required geometry data for almost every object in the
+Solar System.  Other interfaces to _Horizons_ exist (see 'Other Related Work'
+below for several examples), but `lhorizon` is particularly optimized for
+stability and time-to-value for large queries and for processing results with
 reference to arbitrary topocentric coordinate systems.
 
-`lhorizon` offers a flexible, idiomatic, highly performant pseudo-API to _Horizons_ that returns data as standard 
-scientific Python objects (_NumPy_ [@harris2020array] `ndarrays` and _pandas_ [@reback2020pandas] DataFrames). 
-It provides special handling functions for bulk, chunked, and body-listing queries; it also includes an ancillary 
-module, `lhorizon.targeter`, for finding the footprint of an observer’s boresight or field of view in topocentric 
-coordinates on target bodies.  
+`lhorizon` offers a flexible, idiomatic, highly performant pseudo-API
+to _Horizons_ that returns data as standard scientific Python objects (_NumPy_
+[@harris2020array] `ndarrays` and _pandas_ [@reback2020pandas] DataFrames). It
+provides special handling functions for bulk, chunked, and body-listing
+queries; it also includes an ancillary module, `lhorizon.targeter`, for finding
+the footprint of an observer’s boresight or field of view in topocentric
+coordinates on target bodies. 
 
-We wrote `lhorizon` in support of a research effort to use Earth-based radio telescopes, including Arecibo and the Very 
-Large Array (VLA), to perform heat flow mapping of the Moon. We needed to coregister these data with existing models, 
-so we were specifically interested in answering questions like: “in lunar latitude and longitude, where is Arecibo 
-pointing?”. As Earth-based radio telescopes are not primarily designed to answer questions about nearby bodies, their processing 
-pipelines aren't readily suited for such goals as, _e.g._, producing lunar maps. Using these instruments in unusual ways also resulted in additional measurement uncertainties (_e.g._, geometric uncertainty) that we wanted to minimize.  _Horizons_ high-precision and 
-its capability to deliver corrected positions relative to arbitrary topocentric points, was an 
-appealing source of ephemerides. 
+We wrote `lhorizon` in support of a research effort to use Earth-based radio
+telescopes, including Arecibo and the Very Large Array (VLA), to perform heat
+flow mapping of the Moon. We needed to coregister these data with existing
+models, so we were specifically interested in answering questions like: “in
+lunar latitude and longitude, where is Arecibo pointing?”. As Earth-based radio
+telescopes are not primarily designed to answer questions about nearby bodies,
+their processing pipelines aren't readily suited for goals like producing 
+lunar maps. Using these instruments in unusual ways also resulted in
+additional measurement uncertainties (_e.g._, geometric uncertainty) that we
[…]
