### 2021-09-05 MillionConcepts/lhorizon#21 @m-stclair
> I applied most of my suggestions directly to the text in PR #20 (I hope that's convenient for you, else let me know what you prefer).

Thank you for your notes! I've been chasing a deadline for the past week and am just getting back to this now; sorry about the delay. I accepted your PR,  added some rewrites and additional edits in response to your comments, and ticked off the items in your comment I think are probably fully resolved. Further discussion below:

>  **_Horizons_ is the only system in the world that provides...**: not really, other interfaces exist...

You are right; this claim is excessively parochial. I have weakened such statements throughout.

>  **"_Horizons_ is high-precision"**: what does "high precision" mean?

I meant it as a qualitative description, but I have specified it in context. It can be entirely removed if you think it's misleading or irrelevant.

>
>  **"has the capability to produce corrected positions"**: what does "corrected" mean here? "corrected" of what effects/errors/etc?

Clarified this in the text.

>  **"had experienced bitrot"** : this is slang and at least it's not familiar to me

I replaced this term with a full description of the phenomenon.

>
> **"also collecting the appropriate data files, called “kernels.”**: I know by experience that this is often tricky, but wouldn't your examples also get into the same trouble when calling spiceypy...

Most of `lhorizon` does not use SPICE at all, wrapped in `spiceypy` or otherwise. One piece of functionality in an optional portion of the module uses SPICE to do the last step (a reference frame transformation) of an operation that, end-to-end, it is difficult to perform using only either Horizons or SPICE. The value proposition of this piece of functionality should be considered primarily capability rather than usability, although this is not true for most of `lhorizon`. Do you think I need to add an explicit discussion of this somewhere?

> **"automatically-generated kernels"**: what do you mean here?

I meant: there are a number of services that procedurally generate SPICE kernels. _Horizons_, for instance (though only at its telnet endpoint), given various parameters, will custom-produce SPK kernels for small bodies valid in particular time windows, etc. This is really tangential to the paper, though, and I have just removed it.

>  **"up to tens of millions of kernels"**: did you really count them? All version summed together? :) Anyway, the sentence was extremely redundant.

Haha no. Just a rough order-of-magnitude estimate. Anyway, as you say, there is not really any need for this part of the discussion, and I have removed it.

>  as a general remark, I think the paragraph about SPICE was way too harsh and colloquial. Also, it's hard to defend subjective statement such as "extremely high barrier" or "is difficult for planetary scientists and specialized engineers" (it's not, after you get accustomed to it, as for most software packages)...

It is certainly true that assessments of usability have subjective elements, and that training effects are major confounders for these assessments. I think the thrust of your amendments makes sense.

> **"it cover more sites"**: what does this mean? which "sites"?

Clarified this in the text.

>  **"geometry data"**: it would be good to define this somewhere (rather early in the text), as it's somehow the focus and might be unclear to general readers/users. Also, I'm not sure whether "geometry data" or "geometric data" or "geometry" (e.g., of an observation) would be the correct terminology.

Yes, this is an ugly term. I don't know what is best. "Geometry data" is common jargon in NAIF publications, but it might not be used as widely as I think it is. I agree that "geometric" sounds more grammatical, but it is somewhat overloaded in this context, as it is sometimes used to refer to "true" position (as opposed to "astrometric" position with target-observer light-time, and sometimes other corrections, applied). Do you think there is another term that would be clearer to a general audience?

> **Other Related Work**: references to the (very much active) CALCEPH library (https://www.imcce.fr/inpop/calceph/ for C, fortran, python, Matlab interfaces) is missing.

Yes, that's an oversight. I've added this to the text. Do you know if CALCEPH has a preferred citation? I can't find an explicit statement about it.
