### 2021-09-05 MillionConcepts/lhorizon#24 @m-stclair
Many of the tests check against saved csv dumps of output from `Horizons`. However, of course, all of the _inputs_ to the package are also basically live dumps of similar output from `Horizons`, so this is less reassuring than it could be. To guard against the possibility that either the tests or the package as a whole are  parsing `Horizons` output in a pervasively but subtly poor way that is introducing a systematic numerical error, I also have a few black-box "reasonableness" checks, e.g.:

- is the angular diameter of the Moon considerably greater or smaller than I think it should be relative to an Earth observer?
- Are rays drawn from a point on Mars to a point on the Moon, and then back from that point on the Moon to the same point on Mars, reflections of one another to within a small margin of error?

Do you think I should have more of these?  I have certainly considered adding calls to other data sources. Other purely analytical tests could be added, like checking to see if the position `lhorizon` reports for a particular planet is within the expected margin of error from  its approximate position as computed using Keplerian formulae.

In either case, I can add more documentation to `lhorizon.tests`.
