### 2021-09-05 MillionConcepts/lhorizon#23 @m-stclair
> This is an interesting example of application of your package and functions!
> However, am I right to say that "In 8" and the whole Part 2 fully rely on SPICE/spicey-py?

Yes, this specific piece of `lhorizon.target`'s functionality relies on SPICE via spiceypy; unlike most of `lhorizon`, its added value should probably be considered in terms of functionality rather than usability (although if you are only interested in the SPICE built-in frames + the high-precision Moon frames, the included kernels offer a completely abstracted interface). Do you think I need to add a longer discussion of this somewhere? See also my response to your related comment in your thread on the paper.
