### 2021-09-05 MillionConcepts/lhorizon#23 @m-stclair
> Also, minor point: what do you mean by `earth mean equator and equinox of J2000, **often called "J2000" for historical reasons**`? It's simply called J2000 because it refers to the equator at 01/01/2000 12:00:00.

I mean that "J2000" is an ambiguous name for a reference frame. It refers to the time, but there are other reference frames that use J2000 as an epoch but are not the "J2000" frame. I have encountered metadata that referred to quite different frames (like the ecliptic system SPICE calls "ECLIPJ2000") as "J2000"  because they used the J2000 epoch. More narrowly, in this case, I am gesturing to the fact that SPICE now uses the ICRF reference frame rather than the FK5 J2000 frame but retains the "J2000" identifier for it. Because they are so close, it is unlikely to cause numerical problems for the vast majority of use cases, but is still a potential source of confusion -- particularly because _Horizons_ does _not_ retain the "J2000" identifier for this reference frame. For these reasons, I like to give a general warning about this term. If you think it is confusing or unnecessary here, though, I can certainly cut it.
