+++ requirements.txt
@@ -1,0 +1,15 @@
+astroquery~=0.4.3
+jupyter~=1.0.0
+numpy~=1.21.1
+memory_profiler~=0.58.0
+more-itertools~=8.8.0
+pandas~=1.3.1
+pympler~=0.9
+pytest~=6.2.4
+pytest-cov~=2.12.1
+pytest-mock~=3.6.1
+python-dateutil~=2.8.2
+pytz~=2021.1
+requests~=2.26.0
+spiceypy~=4.0.2
+sympy~=1.8
