### 2021-09-05 810afb3
add assert statements for comparison to benchmark notebook
