### 2021-09-05 MillionConcepts/lhorizon#19 @m-stclair
> Love this! Unfortunately the notebook still fails in binder (at least for me, astropy seems to be missing)

Sorry, yes, I hadn't even intended these to work in Binder, because they use (by the standards of the Binder backend) large files and a lot of RAM, and I think it will be difficult to make them consistently not crash Binder. Do you think that adding local installation instructions as @malmans2 suggested is adequate?

> Also, I would suggest to add some `assert` checks among calls results, to make sure/prove that on top of a better timing, one also gets the same results.

Good idea. Added in https://github.com/MillionConcepts/lhorizon/commit/810afb36c83e4137fd4844a3b9d87d8addff2b2f
