### 2021-09-06 18b9050
added extended install and usage instructions for benchmarks
