### 2021-09-06 MillionConcepts/lhorizon#19 @m-stclair
>     (I'm not 100% sure, but I think in this way GitHub will show the README in the benchmark directory without having to click on it)

This appears to have worked!

>      Use step by step instructions in the readme, including terminal commands such as how to grab the files (e.g., wget/curl), how to install the benchmark environment, and how to install lhorizon in the benchmark environment

Added.

>     Consider storing the benchmark files in a different repository/cloud (I believe they are currently in a google drive). E.g., you could use [git-lfs](https://git-lfs.github.com/) and store the file in this GitHub repo, or you could use zenodo.

Ok, I will think about this. Do you mean that you think that the files should have their own DOI? Or instead that it is awkward, or might be overly fragile, to distribute them via Google Drive in this way? git-lfs is basically just an automated aliasing system for external storage locations, right? So this would involve, for instance, setting up an S3 bucket and pointing the git-lfs server at credentials for that bucket (+ some addressing scheme for that bucket) in order to create versioned pointers from this repository to that bucket?

>     I think there is a typo in the environment file. ~`memory-profiler`~ should be `memory_profiler` https://github.com/MillionConcepts/lhorizon/blob/eee7be4ee9e14f72de643bd690163cb0640a657c/benchmark/bench-environment.yml#L16

Yes, that's a typo.

>    I would add `jupyter` in the environment, as it is consistent with the environment file provided in the root directory

Added.
