### 2021-09-06 0012dfe
added explanatory comments and headings to benchmark notebook
