### 2021-09-06 MillionConcepts/lhorizon#19 @m-stclair
>    Although the variable names are self-describing, I would make more obvious which cell is benchmarking lhorizon vs jplhorizon (e.g., add markdowns, comments, or prints)

extra markdowns & comments added in 0012dfe
