### 2021-09-06 30a9140
additional docstrings and comments for tests
