+++ lhorizon/tests/test_lhorizon_utils.py
@@ -1,3 +1,5 @@
+"""unit tests for `lhorizon.utils` functions"""
+
 import math
 import re
 
@@ -34,6 +36,7 @@ def test_hats_2():
 
 
 def test_is_it():
+    """did is_it forget about some important type?"""
     this_range = range(100)
     is_rangelike = is_it(range, list, tuple)
     assert is_rangelike(this_range)
@@ -42,6 +45,7 @@ def test_is_it():
 
 
 def test_sph2cart():
+    """white-box test for sph2cart: does it do trig functions wrong?"""
     up = sph2cart(90, 90, 1)
     assert math.isclose(up[0], 0, abs_tol=1e-9)
     assert math.isclose(up[1], 0, abs_tol=1e-9)
@@ -49,6 +53,11 @@ def test_sph2cart():
 
 
 def test_roundtrip_transform():
+    """
+    black-box test for sph2cart and cart2sph:
+    check that they are strictly inverse, given random coordinates within
+    allowable ranges
+    """
     cart = rng.integers(0, 10, 3)
     assert np.allclose(cart, sph2cart(*cart2sph(*cart)))
     sph = np.array([
@@ -58,6 +67,10 @@ def test_roundtrip_transform():
 
 
 def test_hunt_csv_1():
+    """
+    does hunt_csv parse this goofy but
+    technically-correct delimited text in the way we think it should?
+    """
     raining = re.compile(r"(?<=cats).*(?=dogs)")
     some_csv = "catssomething,something_else,1,2dogs"
     assert hunt_csv(raining, some_csv) == [
@@ -69,6 +82,12 @@ def test_hunt_csv_1():
 
 
 def test_hunt_csv_2():
+    """
+    if we get some random integers and assemble
+    them into a gross little chunk of CSV, then sum each 'line' of the table
+    produced by parsing that chunk with hunt_csv, do we get the same result
+    as summing across a partition of that list of integers?
+    """
     ints = rng.integers(0, 100, 8)
     breaks = rng.choice(("\r", "\n", "\r\n"), 3)
     block = (
@@ -82,6 +101,9 @@ def test_hunt_csv_2():
 
 
 def test_snorm_1():
+    """
+    does snorm constrain this specific interval the way we expect it to?
+    """
     thing = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10]
     normed = snorm(thing, 0, 1, 2, 9)
     assert math.isclose(normed[1], 0)
@@ -89,17 +111,27 @@ def test_snorm_1():
 
 
 def test_snorm_2():
+    """
+    does snorm normalize the center of a
+    randomly-selected normal distribution to something close to 0.5? this test
+    will throw a false-positive failure about one in ten million times.
+    """
     thing = rng.normal(rng.integers(-1000, 1000), rng.integers(1, 100), 100000)
-    # this will fail about one in ten million times.
     assert abs(np.mean(snorm(thing)) - 0.5) < 0.11
 
 
 def test_snorm_3():
+    """
+    does snorm know how to divide things in two?
+    """
     number = rng.integers(-1000, 1000)
     assert math.isclose(snorm(number, 0, 1, 0, number * 2), 0.5)
 
 
 def test_listify():
+    """
+    does listify turn all of these things into lists?
+    """
     assert isinstance(listify("asdfasdf"), list)
     assert isinstance(listify([1, 2, 3, 4]), list)
     assert isinstance(listify(1), list)
