+++ lhorizon/tests/test_targeter_online.py
@@ -1,3 +1,8 @@
+"""
+online tests for lhorizon.target, using responses retrieved live from JPL
+Horizons
+"""
+
 import datetime as dt
 import math
 import warnings
@@ -5,6 +10,9 @@ import warnings
 import numpy as np
 import pandas as pd
 import pytest
+
+# skip these tests if running in an install that doesn't include dependencies
+# for lhorizon.target
 pytest.importorskip("sympy")
 
 from lhorizon import LHorizon
@@ -20,7 +28,17 @@ rng = np.random.default_rng()
 
 
 def test_targeter_long_online():
+    """
+    check "long" (single topocentric point across multiple times)
+    target-finding behavior on live response. this test retrieves angles from
+    JPL Horizons from the geocenter to a fixed point on the lunar surface
+    (10 degrees North, 10 degrees East) at a randomly-selected series of
+    times, then feeds those angles to Targeter and verifies that it calculates
+    targets reasonably close to 10 degrees North, 10 degrees East.
+    """
+    # pick a random time between about 1972 and 2023
     start = rng.integers(2441318, 2460000)
+    # along with the succeeding nine days
     epochs = [start + ix / 10 for ix in range(10)]
     targeter = Targeter(
         LHorizon(301, epochs=epochs), solutions=lunar_solutions
@@ -41,6 +59,15 @@ def test_targeter_long_online():
 
 
 def test_targeter_wide_online():
+    """
+    check "wide" (multiple topocentric points at a single time) target-finding
+    behavior on live response. this test computes IAU_MOON coordinates for
+    rays from geocenter to a coarse grid of ra/dec coordinates across the
+    Moon, then makes a couple of "reasonableness" assertions -- is the target
+    position of a ray to the sublunar point close, in IAU_MOON coordinates,
+    to where Horizons claims the sublunar point currently is?
+    is the apparent angular diameter of the Moon unreasonably large?
+    """
     now = dt.datetime.now().isoformat()
     targeter = Targeter(LHorizon(301, epochs=now), solutions=lunar_solutions)
 

[…]
