### 2021-09-06 MillionConcepts/lhorizon#24 @m-stclair
added some explanatory docstrings and comments to `lhorizon.tests` submodules: 30a9140

leaving this issue open to get your opinion on the broader verification question.
