### 2021-09-06 MillionConcepts/lhorizon#19 @malmans2
I'd personally prefer zenodo because of the DOI or git-lfs because it seems easier to maintain/more robust. But nothing wrong with google drive, so I'll leave it up to you. Regarding git-lfs, you just have to install it, and then set which files to manage through Git LFS. No need to set up an external bucket. See [Getting Started](https://git-lfs.github.com/).
