### 2021-09-06 MillionConcepts/lhorizon#19 @malmans2
Final minor comments:
In point 1: Typo ~Them~ should be Then
In point 2: If you assume that the user is in the root directory the yml path should be `benchmark/bench-environment.yml`. But maybe it's easier to run all commands from `/benchmark`? This way there's no need to move the unzipped data in point 1, and no need to change directory later in point 4.
