### 2021-09-06 MillionConcepts/lhorizon#19 @malmans2
I'd also mention the benchmark directory in the main readme, so users know about it.
