### 2021-09-07 b0ede9f
small documentation changes
