### 2021-09-07 89b1b3c
add references to benchmarks and easier instructions
