### 2021-09-07 MillionConcepts/lhorizon#19 @m-stclair
I always have trouble believing things like git-lfs are free...in any case, I switched over to Zenodo, because using git-lfs will require users to install it to successfully clone the repository, and it seems easier to simply instruct them to download the files. I have also addressed your other comments. All changes are present by 89b1b3c . Leaving this issue open pending response from @steo85it on Binder.
