+++ docs/api.md
@@ -2,6 +2,7 @@
 
 * [\_response\_parsers](#_response_parsers)
   * [make\_lhorizon\_dataframe](#_response_parsers.make_lhorizon_dataframe)
+  * [clean\_visibility\_flags](#_response_parsers.clean_visibility_flags)
   * [clean\_up\_vectors\_series](#_response_parsers.clean_up_vectors_series)
   * [clean\_up\_observer\_series](#_response_parsers.clean_up_observer_series)
   * [clean\_up\_series](#_response_parsers.clean_up_series)
@@ -28,6 +29,7 @@
     * [prepare\_request](#base.LHorizon.prepare_request)
     * [\_\_str\_\_](#base.LHorizon.__str__)
     * [\_\_repr\_\_](#base.LHorizon.__repr__)
+* [benchscratch](#benchscratch)
 * [targeter\_utils](#targeter_utils)
   * [array\_reference\_shift](#targeter_utils.array_reference_shift)
 * [target](#target)
@@ -36,6 +38,7 @@
     * [find\_targets](#target.Targeter.find_targets)
     * [find\_target\_grid](#target.Targeter.find_target_grid)
     * [transform\_targets\_to\_body\_frame](#target.Targeter.transform_targets_to_body_frame)
+* [benchscratch\_2](#benchscratch_2)
 * [handlers](#handlers)
   * [estimate\_line\_count](#handlers.estimate_line_count)
   * [chunk\_time](#handlers.chunk_time)
@@ -68,7 +71,7 @@
 * [kernels.load](#kernels.load)
   * [load\_metakernel](#kernels.load.load_metakernel)
 
-<a name="_response_parsers"></a>
+<a id="_response_parsers"></a>
 
 # Module \_response\_parsers
 
@@ -76,7 +79,7 @@ helper functions for parsing response text from the JPL Horizons CGI.
 these functions are intended to be called by LHorizon methods and should
 generally not be called directly.
 
-<a name="_response_parsers.make_lhorizon_dataframe"></a>
+<a id="_response_parsers.make_lhorizon_dataframe"></a>
 
 #### make\_lhorizon\_dataframe
 
@@ -86,7 +89,19 @@ def make_lhorizon_dataframe(jpl_response: str, topocentric_target: bool = False)
 
 make a DataFrame from Horizons CGI response text.
 
-<a name="_response_parsers.clean_up_vectors_series"></a>
+<a id="_response_parsers.clean_visibility_flags"></a>
+
+#### clean\_visibility\_flags
+
+```python
+def clean_visibility_flags(horizon_dataframe: pd.DataFrame) -> pd.DataFrame
+```
+
+assign names to unlabeled 'visibility flag' columns -- solar presence,
+lunar/interfering body presence, is-target-on-near-side-of-parent-body,
+is-target-illuminated; drop then if empty
+
+<a id="_response_parsers.clean_up_vectors_series"></a>
 
 #### clean\_up\_vectors\_series
 
@@ -96,7 +111,7 @@ def clean_up_vectors_series(pattern: str, series: Array) -> pd.Series
 
 regularize units, format text, and parse dates in a VECTORS table column
 
-<a name="_response_parsers.clean_up_observer_series"></a>
+<a id="_response_parsers.clean_up_observer_series"></a>
 
 #### clean\_up\_observer\_series
 
@@ -106,7 +121,7 @@ def clean_up_observer_series(pattern: str, series: Array) -> Optional[pd.Series]
 
 regularize units, format text, and parse dates in an OBSERVER table column
 
-<a name="_response_parsers.clean_up_series"></a>
+<a id="_response_parsers.clean_up_series"></a>
 
[…]
