### 2021-09-07 5125d80
remove garbage scratch headings from docs
