### 2021-09-07 cf0c99b
squash series of doctest debug commits
