### 2021-09-07 MillionConcepts/lhorizon#14 @m-stclair
doc freshness test added:  cf0c99b
