### 2021-09-07 6b5a947
tweaks to paper and readme text[no actions]
