### 2021-09-07 MillionConcepts/lhorizon#21 @m-stclair
brief discussion of use cases in paper and tweaks to readme: 6b5a947
