### 2021-09-09 MillionConcepts/lhorizon#21 @steo85it
Thanks for your changes and comments, the paper looks clearer now in my opinion. I just suggested a few minor modifications in PR #25 , along with suggesting to use the CALCEPH website as reference for it. The rest is good for me now.
