### 2021-09-09 MillionConcepts/lhorizon#24 @steo85it
Thanks for the detailed explanation: I think the tests are more than adequate and that should now be clear from the docstrings.
