### 2021-09-09 MillionConcepts/lhorizon#19 @steo85it
These solutions works for me, thanks. Just a minor issue (not sure if this is somehow related to my setup): when running the timing comparison cells, I need the `%%time` magic command to be alone on a line of code, even the (currently present) comment causes the cell run to fail with a

> UsageError: Can't use statement directly after '%%time'!

error. Simply sending the comment to the following line (or removing it) solves this.
