### 2021-09-09 MillionConcepts/lhorizon#23 @steo85it
> Yes, this specific piece of `lhorizon.target`'s functionality relies on SPICE via spiceypy; unlike most of `lhorizon`, its added value should probably be considered in terms of functionality rather than usability (although if you are only interested in the SPICE built-in frames + the high-precision Moon frames, the included kernels offer a completely abstracted interface). Do you think I need to add a longer discussion of this somewhere? See also my response to your related comment in your thread on the paper.

I would probably add a comment about this in the notebook (or section) docstring, just mentioning that `for this step, SPICE kernels have been loaded from the XX folder. For applications to different bodies or frames you might need to load different/additional kernels - or different versions thereof.` (or something along these lines).

> I mean that "J2000" is an ambiguous name for a reference frame. It refers to the time, but there are other reference frames that use J2000 as an epoch but are not the "J2000" frame. ...

I would remove the `for historical reasons`, then the sentence would be fine for me.
