### 2021-09-10 4c982dd
fix mangled magics in bench notebook[no actions]
