### 2021-09-10 1fc5646
added comments to target_example.ipynb
