### 2021-09-10 MillionConcepts/lhorizon#23 @m-stclair
> I would probably add a comment about this in the notebook (or section) docstring, just mentioning that `for this step, SPICE kernels have been loaded from the XX folder. For applications to different bodies or frames you might need to load different/additional kernels - or different versions thereof.` (or something along these lines).

I have a comment about that in the section, and added one to the cell containining the call to `load_metakernel()` to make it clearer what that function is doing.

> I would remove the `for historical reasons`, then the sentence would be fine for me.

Got it. I removed that and replaced it with 'earth mean equator and equinox of 01/01/2000 12:00, called "ICRF" by Horizons and "J2000" by SPICE', which is what I really wanted to draw attention to.

These changes are applied in 1fc5646.
