### 2021-09-10 88bf11a
add calceph reference to paper and bibliography[skip ci]
