+++ docs/paper.bib
@@ -140,3 +140,10 @@
     doi          = {10.5281/zenodo.3509134},
     url          = {https://doi.org/10.5281/zenodo.3509134}
 }
+
+@software{calceph,
+    author = {M. Gastineau, J. Laskar, A. Fienga, H. Manche},
+    title = {CALCEPH},
+    publisher = {Institut de mécanique céleste et de calcul des éphémérides},
+    url = {https://www.imcce.fr/inpop/calceph/},
+}
