### 2021-09-10 c70f82d
add date & version to calceph citation[skip ci]
