+++ docs/paper.bib
@@ -144,6 +144,9 @@
 @software{calceph,
     author = {M. Gastineau, J. Laskar, A. Fienga, H. Manche},
     title = {CALCEPH},
+    version = {3.5.0},
+    month = {August},
+    year = {2021},
     publisher = {Institut de mécanique céleste et de calcul des éphémérides},
     url = {https://www.imcce.fr/inpop/calceph/},
 }
