### 2021-09-10 #3495 @malmans2
@arfon all my comments have been addressed by the authors and I think `lhorizon` is suitable for publication!
The package is in line with all best practices for open source software and I'm sure many users will find `lhorizon` very useful!
