### 2021-09-10 #3495 @steo85it
@arfon Same here, all my comments have been addressed and I'd recommend this useful package (and related paper) for publication. Thanks for inviting me to take part to this interesting process!
