### 2021-09-11 27471d2
fix edge case in pd.Series time conversion
