### 2021-09-11 de2fa9c
add url hooks for a couple of VECTORS options
