+++ lhorizon/tests/data/test_cases.py
@@ -27,11 +27,11 @@ TEST_CASES = {
         "%2C40.75%2C0%40499%22&SOLAR_ELONG=%220%2C180%22"
         "&LHA_CUTOFF=0&CSV_FORMAT=YES&CAL_FORMAT=BOTH"
         "&ANG_FORMAT=DEG&APPARENT=AIRLESS&REF_SYSTEM=J2000"
-        "&EXTRA_PREC=NO&CENTER=coord%40399&COORD_TYPE"
-        "=GEODETIC&SITE_COORD=%27-243.460000%2C33.800000%2C0"
-        ".000000%27&START_TIME=%221959-01-01T00%3A00%3A00%22"
-        "&STOP_TIME=%221959-01-02T01%3A00%3A00%22&STEP_SIZE"
-        "=%2230m%22&SKIP_DAYLT=NO",
+        "&EXTRA_PREC=NO&VEC_CORR=%27NONE%27&VEC_TABLE=%273%27"
+        "&CENTER=coord%40399&COORD_TYPE=GEODETIC&SITE_COORD"
+        "=%27-243.460000%2C33.800000%2C0.000000%27&START_TIME"
+        "=%221959-01-01T00%3A00%3A00%22&STOP_TIME=%221959-01"
+        "-02T01%3A00%3A00%22&STEP_SIZE=%2230m%22&SKIP_DAYLT=NO",
         "data_path": str(
             Path(Path(__file__).parent, "CYDONIA_PALM_SPRINGS_1959_TOPO")
         ),
@@ -46,8 +46,8 @@ TEST_CASES = {
         "%2C14%2C15%2C17%2C20%2C45%27&COMMAND=%22301%22"
         "&SOLAR_ELONG=%220%2C180%22&LHA_CUTOFF=0&CSV_FORMAT"
         "=YES&CAL_FORMAT=BOTH&ANG_FORMAT=DEG&APPARENT=AIRLESS"
-        "&REF_SYSTEM=J2000&EXTRA_PREC=NO&CENTER=%275%40399%27"
-        "&TLIST=",
+        "&REF_SYSTEM=J2000&EXTRA_PREC=NO&VEC_CORR=%27NONE%27"
+        "&VEC_TABLE=%273%27&CENTER=%275%40399%27&TLIST=",
     },
     "SUN_PHOBOS_1999": {
         "init_kwargs": {
@@ -60,8 +60,9 @@ TEST_CASES = {
         "%2C14%2C15%2C17%2C20%2C45%27&COMMAND=%22Phobos%22"
         "&SOLAR_ELONG=%220%2C180%22&LHA_CUTOFF=0&CSV_FORMAT"
         "=YES&CAL_FORMAT=BOTH&ANG_FORMAT=DEG&APPARENT=AIRLESS"
-        "&REF_SYSTEM=J2000&EXTRA_PREC=NO&CENTER=%27%40sun%27"
-        "&TLIST=2448257.5&SKIP_DAYLT=NO",
+        "&REF_SYSTEM=J2000&EXTRA_PREC=NO&VEC_CORR=%27NONE%27"
+        "&VEC_TABLE=%273%27&CENTER=%27%40sun%27&TLIST=2448257"
+        ".5&SKIP_DAYLT=NO",
         "data_path": str(Path(Path(__file__).parent, "SUN_PHOBOS_1999")),
     },
     "WEIRD_OPTIONS": {
@@ -89,9 +90,10 @@ TEST_CASES = {
         "+NOFRAG%3B%22&SOLAR_ELONG=%221%2C99%22&LHA_CUTOFF=1"
         "&CSV_FORMAT=YES&CAL_FORMAT=BOTH&ANG_FORMAT=DEG"
[…]
