### 2021-09-12 0398513
Fixing up bibtex entries
