+++ docs/paper.bib
@@ -1,6 +1,6 @@
 % Encoding: UTF-8
 
-@INPROCEEDINGS{2015IAUGA..2256293G,
+@INPROCEEDINGS{Giorgini_2015,
        author = {{Giorgini}, Jon D.},
         title = "{Status of the JPL Horizons Ephemeris System}",
     booktitle = {IAU General Assembly},
@@ -131,7 +131,7 @@
 }
 
 @software{reback2020pandas,
-    author       = {The pandas development team},
+    author       = {{The pandas development team}},
     title        = {pandas-dev/pandas: Pandas},
     month        = feb,
     year         = 2020,
