### 2021-09-12 #3495 @arfon
@m-stclair – could you please merge this PR which fixes a few issues with the current paper: https://github.com/MillionConcepts/lhorizon/pull/26
