+++ setup.py
@@ -2,8 +2,8 @@ from setuptools import setup, find_packages
 
 setup(
     name="lhorizon",
-    version="1.1.0",
-    description="lhorizon helps you find where things are in the solar system",
+    version="1.1.1",
+    description="lhorizon helps you things in the solar system",
     url="https://github.com/millionconcepts/lhorizon.git",
     author="Million Concepts",
     author_email="mstclair@millionconcepts.com",
