### 2021-09-13 #3495 @arfon
> @arfon awesome, thank you! one final request -- I hadn't actually noticed the tags on this issue -- would it be possible for you to remove the HCL / Tcl tags from the article page? GitHub misrecognizes SPICE kernels as code in those languages. If not, I doubt it will cause much confusion.

Yep, that's done now.
